[Web] Adds creation of associations + management of ROLES

+++ Web/src/Insalade/CommunicationBundle/Controller/PostController.php
@@ -164,8 +164,12 @@ class PostController extends Controller
 
         $user = $this->container->get('security.context')->getToken()->getUser();
 
-        if($entity->getCreatorId() != $user->getId() && !$this->get('security.context')->isGranted('ROLE_INSALADE')) {
-            return $this->redirect($this->generateUrl('push_show', array('id' => $id)));
+        if(
+            !$this->get('security.context')->isGranted('ROLE_INSALADE') &&
+            !($this->get('security.context')->isGranted('ROLE_AMICALE') && $entity->getParentId() == $user->getId()) &&
+            !($entity->getCreatorId() == $user->getId())) {
+
+            return $this->redirect($this->generateUrl('post_show', array('id' => $id)));
         }
 
         $editForm = $this->createEditForm($entity);
@@ -210,6 +214,12 @@ class PostController extends Controller
             throw $this->createNotFoundException('Unable to find Post entity.');
         }
 
+        $user = $this->container->get('security.context')->getToken()->getUser();
+
+        if($entity->getCreatorId() != $user->getId() && !$this->get('security.context')->isGranted('ROLE_AMICALE')) {
+            return $this->redirect($this->generateUrl('post_show', array('id' => $id)));
+        }
+
         if($entity->getState() == 'pushed') {
             return $this->redirect($this->generateUrl('post_show', array('id' => $id)));
         }
@@ -244,6 +254,16 @@ class PostController extends Controller
             throw $this->createNotFoundException('Unable to find Post entity.');
         }
 
+        $user = $this->container->get('security.context')->getToken()->getUser();
+
+        if(
+            !$this->get('security.context')->isGranted('ROLE_INSALADE') &&
+            !($this->get('security.context')->isGranted('ROLE_AMICALE') && $entity->getParentId() == $user->getId()) &&
+            !($entity->getCreatorId() == $user->getId())) {
+
+            return $this->redirect($this->generateUrl('post_show', array('id' => $id)));
+        }
+
         $em->remove($entity);
         $em->flush();
 
@@ -258,9 +278,12 @@ class PostController extends Controller
      */
     public function updateStateAction($id, $state)
     {
-        if($this->get('security.context')->isGranted('ROLE_INSALADE')) {
-            $em = $this->getDoctrine()->getManager();
-            $entity = $em->getRepository('InsaladeCommunicationBundle:Post')->find($id);
+        $em = $this->getDoctrine()->getManager();
+        $entity = $em->getRepository('InsaladeCommunicationBundle:Post')->find($id);
+        $user = $this->container->get('security.context')->getToken()->getUser();
+        if(
+            $this->get('security.context')->isGranted('ROLE_INSALADE') OR
+            ($this->get('security.context')->isGranted('ROLE_AMICALE') && $entity->getParentId() == $user->getId())) {
 
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find Post entity.');
